+++ checker/tests/ainfer-testchecker/non-annotated/EnumWithInnerClass.java
@@ -1,21 +1,21 @@
 // This test ensures that enums with inner classes are printed properly to avoid crashing the stub
 // parser, which was a problem with an earlier version of stub-based WPI.
 
-import org.checkerframework.checker.testchecker.ainfer.qual.Sibling1;
+import org.checkerframework.checker.testchecker.ainfer.qual.AinferSibling1;
 
 enum EnumWithInnerClass {
   CONSTANT;
 
   private static class MyInnerClass {
-    int getSibling1() {
-      return (@Sibling1 int) 0;
+    int getAinferSibling1() {
+      return (@AinferSibling1 int) 0;
     }
 
-    void requireSibling1(@Sibling1 int x) {}
+    void requireAinferSibling1(@AinferSibling1 int x) {}
 
     void test() {
       // :: warning: argument
-      requireSibling1(getSibling1());
+      requireAinferSibling1(getAinferSibling1());
     }
   }
 }