+++ checker/tests/ainfer-testchecker/non-annotated/ExistingPurityAnnotations.java
@@ -2,7 +2,7 @@
 // when an astub/ajava/jaif file with existing purity annotations is
 // supplied.
 
-import org.checkerframework.checker.testchecker.ainfer.qual.Sibling1;
+import org.checkerframework.checker.testchecker.ainfer.qual.AinferSibling1;
 import org.checkerframework.framework.qual.EnsuresQualifierIf;
 
 public class ExistingPurityAnnotations {
@@ -14,14 +14,14 @@ public class ExistingPurityAnnotations {
   }
 
   @SuppressWarnings("ainfertest")
-  @EnsuresQualifierIf(expression = "#1", result = true, qualifier = Sibling1.class)
-  public boolean checkSibling1(Object obj1) {
+  @EnsuresQualifierIf(expression = "#1", result = true, qualifier = AinferSibling1.class)
+  public boolean checkAinferSibling1(Object obj1) {
     return true;
   }
 
-  public @Sibling1 Object usePureMethod() {
+  public @AinferSibling1 Object usePureMethod() {
 
-    if (checkSibling1(obj)) {
+    if (checkAinferSibling1(obj)) {
       // If pureMethod doesn't have (and can't infer) an @Pure annotation, this call should
       // unrefine the type of obj, and an error will be issued when
       // we try to return obj on the next line.