+++ checker/tests/ainfer-testchecker/non-annotated/OverriddenMethodsTest.java
@@ -1,13 +1,13 @@
-import org.checkerframework.checker.testchecker.ainfer.qual.Sibling1;
-import org.checkerframework.checker.testchecker.ainfer.qual.Sibling2;
+import org.checkerframework.checker.testchecker.ainfer.qual.AinferSibling1;
+import org.checkerframework.checker.testchecker.ainfer.qual.AinferSibling2;
 
 public class OverriddenMethodsTest {
   static class OverriddenMethodsTestParent {
-    public void foo(@Sibling1 Object obj, @Sibling2 Object obj2) {}
+    public void foo(@AinferSibling1 Object obj, @AinferSibling2 Object obj2) {}
 
-    public void bar(@Sibling1 OverriddenMethodsTestParent this, @Sibling2 Object obj) {}
+    public void bar(@AinferSibling1 OverriddenMethodsTestParent this, @AinferSibling2 Object obj) {}
 
-    public void barz(@Sibling1 OverriddenMethodsTestParent this, @Sibling2 Object obj) {}
+    public void barz(@AinferSibling1 OverriddenMethodsTestParent this, @AinferSibling2 Object obj) {}
 
     public void qux(Object obj1, Object obj2) {
       // :: warning: argument
@@ -24,17 +24,17 @@ public class OverriddenMethodsTest {
     @Override
     public void foo(Object obj, Object obj2) {
       // :: warning: (assignment)
-      @Sibling1 Object o = obj;
+      @AinferSibling1 Object o = obj;
       // :: warning: (assignment)
-      @Sibling2 Object o2 = obj2;
+      @AinferSibling2 Object o2 = obj2;
     }
 
     @Override
     public void bar(Object obj) {
       // :: warning: (assignment)
-      @Sibling1 OverriddenMethodsTestChild child = this;
+      @AinferSibling1 OverriddenMethodsTestChild child = this;
       // :: warning: (assignment)
-      @Sibling2 Object o = obj;
+      @AinferSibling2 Object o = obj;
     }
 
     @SuppressWarnings("all")
@@ -49,7 +49,7 @@ public class OverriddenMethodsTest {
       barz(obj);
     }
 
-    public void callqux(@Sibling1 Object obj1, @Sibling2 Object obj2) {
+    public void callqux(@AinferSibling1 Object obj1, @AinferSibling2 Object obj2) {
       qux(obj1, obj2);
     }
   }