+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/builder/CFGTranslationPhaseOne.java
@@ -231,16 +231,16 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
   final Label exceptionalExitLabel;
 
   /**
-   * Current {@link TryFinallyScopeCell} to which a return statement should jump, or null if there
-   * is no valid destination.
+   * Current {@link LabelCell} to which a return statement should jump, or null if there is no valid
+   * destination.
    */
-  @Nullable TryFinallyScopeCell returnTargetL;
+  @Nullable LabelCell returnTargetLC;
 
   /**
-   * Current {@link TryFinallyScopeCell} to which a break statement with no label should jump, or
-   * null if there is no valid destination.
+   * Current {@link LabelCell} to which a break statement with no label should jump, or null if
+   * there is no valid destination.
    */
-  @Nullable TryFinallyScopeCell breakTargetL;
+  @Nullable LabelCell breakTargetLC;
 
   /**
    * Map from AST label Names to CFG {@link Label}s for breaks. Each labeled statement creates two
@@ -249,10 +249,10 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
   Map<Name, Label> breakLabels;
 
   /**
-   * Current {@link TryFinallyScopeCell} to which a continue statement with no label should jump, or
-   * null if there is no valid destination.
+   * Current {@link LabelCell} to which a continue statement with no label should jump, or null if
+   * there is no valid destination.
    */
-  @Nullable TryFinallyScopeCell continueTargetL;
+  @Nullable LabelCell continueTargetLC;
 
   /**
    * Map from AST label Names to CFG {@link Label}s for continues. Each labeled statement creates
@@ -399,7 +399,7 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     regularExitLabel = new Label();
     exceptionalExitLabel = new Label();
     tryStack = new TryStack(exceptionalExitLabel);
-    returnTargetL = new TryFinallyScopeCell(regularExitLabel);
+    returnTargetLC = new LabelCell(regularExitLabel);
     breakLabels = new HashMap<>(2);
     continueLabels = new HashMap<>(2);
     returnNodes = new ArrayList<>();
@@ -816,9 +816,15 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     }
   }
 
-  /** Add the label {@code l} to the extended node that will be placed next in the sequence. */
+  /**
+   * Add the label {@code l} to the extended node that will be placed next in the sequence.
+   *
+   * @param l the node to add to the forthcoming extended node
+   */
   protected void addLabelForNextNode(Label l) {
-    assert !bindings.containsKey(l);
+    if (bindings.containsKey(l)) {
+      throw new BugInCF("bindings already contains key %s: %s", l, bindings);
+    }
     leaders.add(nodeList.size());
     bindings.put(l, nodeList.size());
   }
@@ -1430,7 +1436,7 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     }
 
     List<Node> arguments;
-    if (TreeUtils.isEnumSuper(tree)) {
+    if (TreeUtils.isEnumSuperCall(tree)) {
       // Don't convert arguments for enum super calls.  The AST contains no actual arguments, while
       // the method element expects two arguments, leading to an exception in convertCallArguments.
       // Since no actual arguments are present in the AST that is being checked, it shouldn't cause
@@ -2161,28 +2167,28 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
           // see JLS 15.23 and 15.24
 
           // all necessary labels
-          Label rightStartL = new Label();
-          Label shortCircuitL = new Label();
+          Label rightStartLabel = new Label();
+          Label shortCircuitLabel = new Label();
 
           // left-hand side
           Node left = scan(leftTree, p);
 
           ConditionalJump cjump;
           if (kind == Tree.Kind.CONDITIONAL_AND) {
-            cjump = new ConditionalJump(rightStartL, shortCircuitL);
+            cjump = new ConditionalJump(rightStartLabel, shortCircuitLabel);
             cjump.setFalseFlowRule(FlowRule.ELSE_TO_ELSE);
           } else {
-            cjump = new ConditionalJump(shortCircuitL, rightStartL);
+            cjump = new ConditionalJump(shortCircuitLabel, rightStartLabel);
             cjump.setTrueFlowRule(FlowRule.THEN_TO_THEN);
           }
           extendWithExtendedNode(cjump);
 
           // right-hand side
-          addLabelForNextNode(rightStartL);
+          addLabelForNextNode(rightStartLabel);
           Node right = scan(rightTree, p);
 
           // conditional expression itself
-          addLabelForNextNode(shortCircuitL);
+          addLabelForNextNode(shortCircuitLabel);
           Node node;
           if (kind == Tree.Kind.CONDITIONAL_AND) {
             node = new ConditionalAndNode(tree, left, right);
@@ -2212,9 +2218,9 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
   public Node visitBreak(BreakTree tree, Void p) {
     Name label = tree.getLabel();
     if (label == null) {
-      assert breakTargetL != null : "no target for break statement";
+      assert breakTargetLC != null : "no target for break statement";
 
-      extendWithExtendedNode(new UnconditionalJump(breakTargetL.accessLabel()));
+      extendWithExtendedNode(new UnconditionalJump(breakTargetLC.accessLabel()));
     } else {
       assert breakLabels.containsKey(label);
 
@@ -2239,8 +2245,8 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
   private class SwitchBuilder {
 
     /**
-     * The tree for the switch statement or switch expression. Its type may be {@link SwitchTree} or
-     * {@code SwitchExpressionTree}}
+     * The tree for the switch statement or switch expression. Its type may be {@link SwitchTree}
+     * (for a switch statement) or {@code SwitchExpressionTree}.
      */
     private final Tree switchTree;
 
@@ -2251,7 +2257,7 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
      * The Tree for the selector expression.
      *
      * <pre>
-     *   switch ( <em> selector expression</em> ) { ... }
+     *   switch ( <em>selector expression</em> ) { ... }
      * </pre>
      */
     private final ExpressionTree selectorExprTree;
@@ -2265,8 +2271,8 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     private AssignmentNode selectorExprAssignment;
 
     /**
-     * If {@link #switchTree} is a switch expression, then this is the synthetic variable tree that
-     * all results of {@code #switchTree} are assigned. Otherwise, this is null.
+     * If {@link #switchTree} is a switch expression, then this is a result variable: the synthetic
+     * variable that all results of {@code #switchTree} are assigned to. Otherwise, this is null.
      */
     private @Nullable VariableTree switchExprVarTree;
 
@@ -2277,7 +2283,7 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
      */
     private SwitchBuilder(Tree switchTree) {
       this.switchTree = switchTree;
-      if (switchTree instanceof SwitchTree) {
+      if (TreeUtils.isSwitchStatement(switchTree)) {
         SwitchTree switchStatementTree = (SwitchTree) switchTree;
         this.caseTrees = switchStatementTree.getCases();
         this.selectorExprTree = switchStatementTree.getExpression();
@@ -2299,19 +2305,20 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     public @Nullable SwitchExpressionNode build() {
       SwitchBuilder oldSwitchBuilder = switchBuilder;
       switchBuilder = this;
-      TryFinallyScopeCell oldBreakTargetL = breakTargetL;
-      breakTargetL = new TryFinallyScopeCell(new Label());
-      int cases = caseBodyLabels.length - 1;
-      for (int i = 0; i < cases; ++i) {
+      LabelCell oldBreakTargetLC = breakTargetLC;
+      breakTargetLC = new LabelCell(new Label());
+      int numCases = caseTrees.size();
+
+      for (int i = 0; i < numCases; ++i) {
         caseBodyLabels[i] = new Label();
       }
-      caseBodyLabels[cases] = breakTargetL.peekLabel();
+      caseBodyLabels[numCases] = breakTargetLC.peekLabel();
 
       buildSelector();
 
       buildSwitchExpressionVar();
 
-      if (switchTree.getKind() == Tree.Kind.SWITCH) {
+      if (TreeUtils.isSwitchStatement(switchTree)) {
         // It's a switch statement, not a switch expression.
         extendWithNode(
             new MarkerNode(
@@ -2321,25 +2328,25 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       }
 
       // Build CFG for the cases.
-      Integer defaultIndex = null;
-      for (int i = 0; i < cases; ++i) {
+      int defaultIndex = -1;
+      for (int i = 0; i < numCases; ++i) {
         CaseTree caseTree = caseTrees.get(i);
-        if (TreeUtils.caseTreeGetExpressions(caseTree).isEmpty()) {
+        if (TreeUtils.isDefaultCaseTree(caseTree)) {
           defaultIndex = i;
         } else {
           buildCase(caseTree, i);
         }
       }
-      if (defaultIndex != null) {
+      if (defaultIndex != -1) {
         // The checks of all cases must happen before the default case, therefore we build the
         // default case last.
         // Fallthrough is still handled correctly with the caseBodyLabels.
         buildCase(caseTrees.get(defaultIndex), defaultIndex);
       }
 
-      addLabelForNextNode(breakTargetL.peekLabel());
-      breakTargetL = oldBreakTargetL;
-      if (switchTree.getKind() == Tree.Kind.SWITCH) {
+      addLabelForNextNode(breakTargetLC.peekLabel());
+      breakTargetLC = oldBreakTargetLC;
+      if (TreeUtils.isSwitchStatement(switchTree)) {
         // It's a switch statement, not a switch expression.
         extendWithNode(
             new MarkerNode(
@@ -2349,7 +2356,7 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       }
 
       switchBuilder = oldSwitchBuilder;
-      if (switchTree.getKind() != Tree.Kind.SWITCH) {
+      if (!TreeUtils.isSwitchStatement(switchTree)) {
         // It's a switch expression, not a switch statement.
         IdentifierTree switchExprVarUseTree = treeBuilder.buildVariableUse(switchExprVarTree);
         handleArtificialTree(switchExprVarUseTree);
@@ -2406,7 +2413,7 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
      * whose value is the value of the switch expression.
      */
     private void buildSwitchExpressionVar() {
-      if (switchTree.getKind() == Tree.Kind.SWITCH) {
+      if (TreeUtils.isSwitchStatement(switchTree)) {
         // A switch statement does not have a value, so do nothing.
         return;
       }
@@ -2422,49 +2429,61 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     }
 
     /**
-     * Build the CFG for the case tree, {@code tree}.
+     * Build the CFG for the given case tree.
      *
-     * @param tree a case tree whose CFG is built
+     * @param tree a case tree whose CFG to build
      * @param index the index of the case tree in {@link #caseBodyLabels}
      */
     private void buildCase(CaseTree tree, int index) {
-      final Label thisBodyL = caseBodyLabels[index];
-      final Label nextBodyL = caseBodyLabels[index + 1];
-      final Label nextCaseL = new Label();
+      boolean isDefaultCase = TreeUtils.isDefaultCaseTree(tree);
 
-      List<? extends ExpressionTree> exprTrees = TreeUtils.caseTreeGetExpressions(tree);
-      if (!exprTrees.isEmpty()) {
-        // non-default cases
+      final Label thisBodyLabel = caseBodyLabels[index];
+      final Label nextBodyLabel = caseBodyLabels[index + 1];
+      final Label nextCaseLabel = new Label();
+
+      // Handle the case expressions
+      if (!isDefaultCase) {
+        // non-default cases: a case expression exists
         ArrayList<Node> exprs = new ArrayList<>();
-        for (ExpressionTree exprTree : exprTrees) {
+        for (ExpressionTree exprTree : TreeUtils.caseTreeGetExpressions(tree)) {
           exprs.add(scan(exprTree, null));
         }
         CaseNode test = new CaseNode(tree, selectorExprAssignment, exprs, env.getTypeUtils());
         extendWithNode(test);
-        extendWithExtendedNode(new ConditionalJump(thisBodyL, nextCaseL));
+        extendWithExtendedNode(new ConditionalJump(thisBodyLabel, nextCaseLabel));
       }
-      addLabelForNextNode(thisBodyL);
+
+      // Handle the case body
+      addLabelForNextNode(thisBodyLabel);
       if (tree.getStatements() != null) {
-        // This is a switch labeled statement groups.
+        // This is a switch labeled statement group.
+        // A "switch labeled statement group" is a "case L:" label along with its code.
+        // The code either ends with a "yield" statement, or it falls through.
         for (StatementTree stmt : tree.getStatements()) {
           scan(stmt, null);
         }
-        // Handle possible fall through by adding jump to next body.
-        extendWithExtendedNode(new UnconditionalJump(nextBodyL));
+        // Handle possible fallthrough by adding jump to next body.
+        if (!isDefaultCase) {
+          extendWithExtendedNode(new UnconditionalJump(nextBodyLabel));
+        }
       } else {
-        // This is a switch rule.
+        // This is either the default case or a switch labeled rule (which appears in a switch
+        // expression).
+        // A "switch labeled rule" is a "case L ->" label along with its code.
         Tree bodyTree = TreeUtils.caseTreeGetBody(tree);
-        if (switchTree.getKind() != Tree.Kind.SWITCH && bodyTree instanceof ExpressionTree) {
+        if (!TreeUtils.isSwitchStatement(switchTree) && bodyTree instanceof ExpressionTree) {
           buildSwitchExpressionResult((ExpressionTree) bodyTree);
         } else {
           scan(bodyTree, null);
           // Switch rules never fall through so add jump to the break target.
-          assert breakTargetL != null : "no target for case statement";
-          extendWithExtendedNode(new UnconditionalJump(breakTargetL.accessLabel()));
+          assert breakTargetLC != null : "no target for case statement";
+          extendWithExtendedNode(new UnconditionalJump(breakTargetLC.accessLabel()));
         }
       }
 
-      addLabelForNextNode(nextCaseL);
+      if (!isDefaultCase) {
+        addLabelForNextNode(nextCaseLabel);
+      }
     }
 
     /**
@@ -2475,7 +2494,8 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
      *   <li>Builds the CFG for the switch expression result.
      *   <li>Creates an assignment node for the assignment of {@code resultExpression} to {@code
      *       switchExprVarTree}.
-     *   <li>Adds an unconditional jump to {@link #breakTargetL} (the end of the switch expression).
+     *   <li>Adds an unconditional jump to {@link #breakTargetLC} (the end of the switch
+     *       expression).
      * </ol>
      *
      * @param resultExpression the result of a switch expression; either from a yield or an
@@ -2499,8 +2519,8 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       assignmentNode.setInSource(false);
       extendWithNode(assignmentNode);
       // Switch rules never fall through so add jump to the break target.
-      assert breakTargetL != null : "no target for case statement";
-      extendWithExtendedNode(new UnconditionalJump(breakTargetL.accessLabel()));
+      assert breakTargetLC != null : "no target for case statement";
+      extendWithExtendedNode(new UnconditionalJump(breakTargetLC.accessLabel()));
     }
   }
 
@@ -2619,9 +2639,9 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
   public Node visitContinue(ContinueTree tree, Void p) {
     Name label = tree.getLabel();
     if (label == null) {
-      assert continueTargetL != null : "no target for continue statement";
+      assert continueTargetLC != null : "no target for continue statement";
 
-      extendWithExtendedNode(new UnconditionalJump(continueTargetL.accessLabel()));
+      extendWithExtendedNode(new UnconditionalJump(continueTargetLC.accessLabel()));
     } else {
       assert continueLabels.containsKey(label);
 
@@ -2647,11 +2667,11 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       conditionStart = new Label();
     }
 
-    TryFinallyScopeCell oldBreakTargetL = breakTargetL;
-    breakTargetL = new TryFinallyScopeCell(loopExit);
+    LabelCell oldBreakTargetLC = breakTargetLC;
+    breakTargetLC = new LabelCell(loopExit);
 
-    TryFinallyScopeCell oldContinueTargetL = continueTargetL;
-    continueTargetL = new TryFinallyScopeCell(conditionStart);
+    LabelCell oldContinueTargetLC = continueTargetLC;
+    continueTargetLC = new LabelCell(conditionStart);
 
     // Loop body
     addLabelForNextNode(loopEntry);
@@ -2668,8 +2688,8 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     // Loop exit
     addLabelForNextNode(loopExit);
 
-    breakTargetL = oldBreakTargetL;
-    continueTargetL = oldContinueTargetL;
+    breakTargetLC = oldBreakTargetLC;
+    continueTargetLC = oldContinueTargetLC;
 
     return null;
   }
@@ -2702,11 +2722,11 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       updateStart = new Label();
     }
 
-    TryFinallyScopeCell oldBreakTargetL = breakTargetL;
-    breakTargetL = new TryFinallyScopeCell(loopExit);
+    LabelCell oldBreakTargetLC = breakTargetLC;
+    breakTargetLC = new LabelCell(loopExit);
 
-    TryFinallyScopeCell oldContinueTargetL = continueTargetL;
-    continueTargetL = new TryFinallyScopeCell(updateStart);
+    LabelCell oldContinueTargetLC = continueTargetLC;
+    continueTargetLC = new LabelCell(updateStart);
 
     // Distinguish loops over Iterables from loops over arrays.
 
@@ -2956,8 +2976,8 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     // Loop exit
     addLabelForNextNode(loopExit);
 
-    breakTargetL = oldBreakTargetL;
-    continueTargetL = oldContinueTargetL;
+    breakTargetLC = oldBreakTargetLC;
+    continueTargetLC = oldContinueTargetLC;
 
     return null;
   }
@@ -3004,11 +3024,11 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       updateStart = new Label();
     }
 
-    TryFinallyScopeCell oldBreakTargetL = breakTargetL;
-    breakTargetL = new TryFinallyScopeCell(loopExit);
+    LabelCell oldBreakTargetLC = breakTargetLC;
+    breakTargetLC = new LabelCell(loopExit);
 
-    TryFinallyScopeCell oldContinueTargetL = continueTargetL;
-    continueTargetL = new TryFinallyScopeCell(updateStart);
+    LabelCell oldContinueTargetLC = continueTargetLC;
+    continueTargetLC = new LabelCell(updateStart);
 
     // Initializer
     for (StatementTree init : tree.getInitializer()) {
@@ -3039,8 +3059,8 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     // Loop exit
     addLabelForNextNode(loopExit);
 
-    breakTargetL = oldBreakTargetL;
-    continueTargetL = oldContinueTargetL;
+    breakTargetLC = oldBreakTargetLC;
+    continueTargetLC = oldContinueTargetLC;
 
     return null;
   }
@@ -3147,15 +3167,15 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     // nodes. Labeled loops must look up and use the continue Labels.
     Name labelName = tree.getLabel();
 
-    Label breakL = new Label(labelName + "_break");
-    Label continueL = new Label(labelName + "_continue");
+    Label breakLabel = new Label(labelName + "_break");
+    Label continueLabel = new Label(labelName + "_continue");
 
-    breakLabels.put(labelName, breakL);
-    continueLabels.put(labelName, continueL);
+    breakLabels.put(labelName, breakLabel);
+    continueLabels.put(labelName, continueLabel);
 
     scan(tree.getStatement(), p);
 
-    addLabelForNextNode(breakL);
+    addLabelForNextNode(breakLabel);
 
     breakLabels.remove(labelName);
     continueLabels.remove(labelName);
@@ -3304,7 +3324,7 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       extendWithNode(result);
     }
 
-    extendWithExtendedNode(new UnconditionalJump(this.returnTargetL.accessLabel()));
+    extendWithExtendedNode(new UnconditionalJump(this.returnTargetLC.accessLabel()));
 
     return result;
   }
@@ -3396,10 +3416,10 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
             catches);
 
     // Store return/break/continue labels, just in case we need them for a finally block.
-    TryFinallyScopeCell oldReturnTargetL = returnTargetL;
-    TryFinallyScopeCell oldBreakTargetL = breakTargetL;
+    LabelCell oldReturnTargetLC = returnTargetLC;
+    LabelCell oldBreakTargetLC = breakTargetLC;
     Map<Name, Label> oldBreakLabels = breakLabels;
-    TryFinallyScopeCell oldContinueTargetL = continueTargetL;
+    LabelCell oldContinueTargetLC = continueTargetLC;
     Map<Name, Label> oldContinueLabels = continueLabels;
 
     Label finallyLabel = null;
@@ -3411,12 +3431,12 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       exceptionalFinallyLabel = new Label();
       tryStack.pushFrame(new TryFinallyFrame(exceptionalFinallyLabel));
 
-      returnTargetL = new TryFinallyScopeCell();
+      returnTargetLC = new LabelCell();
 
-      breakTargetL = new TryFinallyScopeCell();
+      breakTargetLC = new LabelCell();
       breakLabels = new TryFinallyScopeMap();
 
-      continueTargetL = new TryFinallyScopeCell();
+      continueTargetLC = new LabelCell();
       continueLabels = new TryFinallyScopeMap();
     }
 
@@ -3515,9 +3535,9 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
         throwing.setTerminatesExecution(true);
       }
 
-      if (returnTargetL.wasAccessed()) {
-        addLabelForNextNode(returnTargetL.peekLabel());
-        returnTargetL = oldReturnTargetL;
+      if (returnTargetLC.wasAccessed()) {
+        addLabelForNextNode(returnTargetLC.peekLabel());
+        returnTargetLC = oldReturnTargetLC;
 
         extendWithNode(
             new MarkerNode(
@@ -3530,14 +3550,14 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
                 tree,
                 "end of finally block for return #" + TreeUtils.treeUids.get(tree),
                 env.getTypeUtils()));
-        extendWithExtendedNode(new UnconditionalJump(returnTargetL.accessLabel()));
+        extendWithExtendedNode(new UnconditionalJump(returnTargetLC.accessLabel()));
       } else {
-        returnTargetL = oldReturnTargetL;
+        returnTargetLC = oldReturnTargetLC;
       }
 
-      if (breakTargetL.wasAccessed()) {
-        addLabelForNextNode(breakTargetL.peekLabel());
-        breakTargetL = oldBreakTargetL;
+      if (breakTargetLC.wasAccessed()) {
+        addLabelForNextNode(breakTargetLC.peekLabel());
+        breakTargetLC = oldBreakTargetLC;
 
         extendWithNode(
             new MarkerNode(
@@ -3550,9 +3570,9 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
                 tree,
                 "end of finally block for break #" + TreeUtils.treeUids.get(tree),
                 env.getTypeUtils()));
-        extendWithExtendedNode(new UnconditionalJump(breakTargetL.accessLabel()));
+        extendWithExtendedNode(new UnconditionalJump(breakTargetLC.accessLabel()));
       } else {
-        breakTargetL = oldBreakTargetL;
+        breakTargetLC = oldBreakTargetLC;
       }
 
       Map<Name, Label> accessedBreakLabels = ((TryFinallyScopeMap) breakLabels).getAccessedNames();
@@ -3584,9 +3604,9 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
         breakLabels = oldBreakLabels;
       }
 
-      if (continueTargetL.wasAccessed()) {
-        addLabelForNextNode(continueTargetL.peekLabel());
-        continueTargetL = oldContinueTargetL;
+      if (continueTargetLC.wasAccessed()) {
+        addLabelForNextNode(continueTargetLC.peekLabel());
+        continueTargetLC = oldContinueTargetLC;
 
         extendWithNode(
             new MarkerNode(
@@ -3599,9 +3619,9 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
                 tree,
                 "end of finally block for continue #" + TreeUtils.treeUids.get(tree),
                 env.getTypeUtils()));
-        extendWithExtendedNode(new UnconditionalJump(continueTargetL.accessLabel()));
+        extendWithExtendedNode(new UnconditionalJump(continueTargetLC.accessLabel()));
       } else {
-        continueTargetL = oldContinueTargetL;
+        continueTargetLC = oldContinueTargetLC;
       }
 
       Map<Name, Label> accessedContinueLabels =
@@ -3934,11 +3954,11 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
       conditionStart = new Label();
     }
 
-    TryFinallyScopeCell oldBreakTargetL = breakTargetL;
-    breakTargetL = new TryFinallyScopeCell(loopExit);
+    LabelCell oldBreakTargetLC = breakTargetLC;
+    breakTargetLC = new LabelCell(loopExit);
 
-    TryFinallyScopeCell oldContinueTargetL = continueTargetL;
-    continueTargetL = new TryFinallyScopeCell(conditionStart);
+    LabelCell oldContinueTargetLC = continueTargetLC;
+    continueTargetLC = new LabelCell(conditionStart);
 
     // Condition
     addLabelForNextNode(conditionStart);
@@ -3972,8 +3992,8 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     // Loop exit
     addLabelForNextNode(loopExit);
 
-    breakTargetL = oldBreakTargetL;
-    continueTargetL = oldContinueTargetL;
+    breakTargetLC = oldBreakTargetLC;
+    continueTargetLC = oldContinueTargetLC;
 
     return null;
   }