+++ framework/src/test/java/org/checkerframework/framework/testchecker/lubglb/quals/PolyLubglb.java
@@ -5,13 +5,10 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import org.checkerframework.framework.qual.DefaultFor;
-import org.checkerframework.framework.qual.SubtypeOf;
-import org.checkerframework.framework.qual.TypeUseLocation;
+import org.checkerframework.framework.qual.PolymorphicQualifier;
 
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
-@SubtypeOf({D.class, E.class})
-@DefaultFor({TypeUseLocation.LOWER_BOUND})
-public @interface F {}
+@PolymorphicQualifier(LubglbA.class)
+public @interface PolyLubglb {}