+++ javacutil/src/main/java/org/checkerframework/javacutil/TreeUtils.java
@@ -1533,7 +1533,7 @@ public final class TreeUtils {
    * @param tree the method invocation to check
    * @return true if this is a super call to the {@link Enum} constructor
    */
-  public static boolean isEnumSuper(MethodInvocationTree tree) {
+  public static boolean isEnumSuperCall(MethodInvocationTree tree) {
     ExecutableElement ex = TreeUtils.elementFromUse(tree);
     assert ex != null : "@AssumeAssertion(nullness): tree kind";
     Name name = ElementUtils.getQualifiedClassName(ex);
@@ -2045,8 +2045,19 @@ public final class TreeUtils {
   }
 
   /**
-   * Get the list of expressions from a case expression. In JDK 11 and earlier, this is a singleton
-   * list. In JDK 12 onwards, there can be multiple expressions per case.
+   * Returns true if this is the default case for a switch statement or expression.
+   *
+   * @param caseTree a case tree
+   * @return true if {@code caseTree} is the default case for a switch statement or expression
+   */
+  public static boolean isDefaultCaseTree(CaseTree caseTree) {
+    return caseTreeGetExpressions(caseTree).isEmpty();
+  }
+
+  /**
+   * Get the list of expressions from a case expression. For the default case, this is empty.
+   * Otherwise, in JDK 11 and earlier, this is a singleton list. In JDK 12 onwards, there can be
+   * multiple expressions per case.
    *
    * @param caseTree the case expression to get the expressions from
    * @return the list of expressions in the case
@@ -2199,6 +2210,16 @@ public final class TreeUtils {
           "TreeUtils.switchExpressionTreeGetCases: reflection failed for tree: %s",
           switchExpressionTree, e);
     }
+  }
+
+  /**
+   * Returns true if the given tree is a switch statement (as opposed to a switch expression).
+   *
+   * @param tree the switch statement or expression to check
+   * @return true if the given tree is a switch statement (as opposed to a switch expression)
+   */
+  public static boolean isSwitchStatement(Tree tree) {
+    return tree.getKind() == Tree.Kind.SWITCH;
   }
 
   /**